Add an optional Essentials add-on for the native agent
Without it a plain .NET app gets 501 not_supported for preferences, secure
storage, device info, display, battery, connectivity, permissions, geolocation
and sensors. Essentials turns out to be usable on its own -- it needs the MAUI
workload but not Controls -- so those endpoints can be lit up without dragging
the app into MAUI.

The endpoint bodies are not copied. They were lifted out of
MauiDevFlowAgentService into src/DevFlow/Shared.Essentials, which both the MAUI
agent and the add-on compile, so the two agents cannot drift. Both services now
hold an EssentialsAgentSupport and delegate. Platform error shaping moved from
protected to public on DevFlowAgentService for the same reason: it is protocol,
and a non-subclass backend has to be able to emit it byte-identically.

Theme stays unsupported on native -- MAUI's theme endpoints read
Application.RequestedTheme, which has no equivalent outside Controls.

Wiring the add-on is explicit (EssentialsDevFlowAgent.Start) rather than
reflection-discovered, so it survives the trimming and AOT that plain .NET iOS
apps depend on. DevFlowAgent.Start gained a factory overload for that.

CI now builds the four native sample heads on macOS via a new solution filter;
they are the executable proof the agent works with no MAUI, and were previously
compiled by nothing.

Copilot-Session: 168e7050-987f-4782-9527-a8fa40e67bdd
