+++ src/DevFlow/DevFlow.slnf
@@ -5,6 +5,7 @@
       "src\\DevFlow\\Microsoft.Maui.DevFlow.Agent.Abstractions\\Microsoft.Maui.DevFlow.Agent.Abstractions.csproj",
       "src\\DevFlow\\Microsoft.Maui.DevFlow.Agent.Core\\Microsoft.Maui.DevFlow.Agent.Core.csproj",
       "src\\DevFlow\\Microsoft.Maui.DevFlow.Agent.Native\\Microsoft.Maui.DevFlow.Agent.Native.csproj",
+      "src\\DevFlow\\Microsoft.Maui.DevFlow.Agent.Native.Essentials\\Microsoft.Maui.DevFlow.Agent.Native.Essentials.csproj",
       "src\\DevFlow\\Microsoft.Maui.DevFlow.Analyzers\\Microsoft.Maui.DevFlow.Analyzers.csproj",
       "src\\DevFlow\\Microsoft.Maui.DevFlow.Agent\\Microsoft.Maui.DevFlow.Agent.csproj",
       "src\\DevFlow\\Microsoft.Maui.DevFlow.Agent.Gtk\\Microsoft.Maui.DevFlow.Agent.Gtk.csproj",
