+++ src/DevFlow/Shared.Essentials/SensorManager.cs
